+++ CHANGES.rst
@@ -13,6 +13,8 @@ New Features
 
 - The ``specviz.load_spectrum`` method is deprecated; use ``specviz.load_data`` instead. [#2273]
 
+- Add first-pass launcher to select config and auto-identify data. [#2257]
+
 Cubeviz
 ^^^^^^^